requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=2.0.0
websockets>=10.0
aiohttp>=3.8.0
python-jose>=3.3.0
cryptography>=3.4.0
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0 